--- src/main/java/com/nhnacademy/book/Image.java
@@ -1,23 +1,0 @@
-package com.nhnacademy.book;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Image {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "image_id")
-    private long imageId;
-
-    @Column(name = "path", nullable = false)
-    private String imagePath;
-}
