+++ src/main/java/com/nhnacademy/book/member/domain/StatusName.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.book;
+package com.nhnacademy.book.member.domain;
 
 public enum StatusName {
     ACTIVE,
